start expanding zlib interface

+++ src/isal/crc.pxd
@@ -24,5 +24,5 @@ cdef extern from "<isa-l/crc.h>":
     cdef unsigned long crc32_gzip_refl(
     unsigned long init_crc,          #!< initial CRC value, 32 bits
     const unsigned char *buf, #!< buffer to calculate CRC on
-    unsigned long long len                #!< buffer length in bytes (64-bit data)
+    Py_ssize_t len                #!< buffer length in bytes (64-bit data)
     )
